+++ stack/create.c
@@ -6,13 +6,13 @@
 /*   By: akarafi <akarafi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/22 16:49:47 by akarafi           #+#    #+#             */
-/*   Updated: 2021/12/22 16:50:38 by akarafi          ###   ########.fr       */
+/*   Updated: 2021/12/23 19:17:39 by akarafi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../push_swap.h"
 
-t_node	*new_node(long n)
+t_node	*new_node(int n)
 {
 	t_node	*node;
